avoid float->double->float by using float version of math functions
This saves a conversion instruction before and after the function call or inline instruction, but more importantly several of the functions are substantially faster in their float form than double form.  Q&D benchmarking shows that:
   sinf/cosf run in about 1/2 the time of sin/cos
   powf runs in about 3/5 the time of pow
   expf runs in about 3/4 the time of exp

Square root compiles to a single instruction, so sqrt and sqrtf take the same amount of time, but we still save the two conversions.

+++ src/common/gaussian.c
@@ -37,8 +37,8 @@ static void compute_gauss_params(const float sigma, dt_gaussian_order_t order, f
                                  float *a2, float *a3, float *b1, float *b2, float *coefp, float *coefn)
 {
   const float alpha = 1.695f / sigma;
-  const float ema = exp(-alpha);
-  const float ema2 = exp(-2.0f * alpha);
+  const float ema = expf(-alpha);
+  const float ema2 = expf(-2.0f * alpha);
   *b1 = -2.0f * ema;
   *b2 = ema2;
   *a0 = 0.0f;
